+++ src/app/[locale]/(private)/module/studysheet/[id]/components/external-materials-table.tsx
@@ -3,15 +3,15 @@ import React from 'react';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
 import { ProfilePicture } from '@/components/ui/profile-picture';
 import { Link } from '@/i18n/routing';
+import { ExternalMaterials } from '@/types/models/current-control/materials';
 import { useTranslations } from 'next-intl';
-import { InternalMaterials } from '@/types/models/current-control/materials';
 
-interface InternalMaterialsTableProps {
-  internalMaterials: InternalMaterials[];
+interface ExternalMaterialsTableProps {
+  externalMaterials: ExternalMaterials[];
   t: ReturnType<typeof useTranslations>;
 }
 
-export const InternalMaterialsTable: React.FC<InternalMaterialsTableProps> = ({ internalMaterials, t }) => {
+export const ExternalMaterialsTable: React.FC<ExternalMaterialsTableProps> = ({ externalMaterials, t }) => {
   return (
     <Table>
       <TableHeader>
@@ -22,13 +22,11 @@ export const InternalMaterialsTable: React.FC<InternalMaterialsTableProps> = ({ 
         </TableRow>
       </TableHeader>
       <TableBody>
-        {internalMaterials.map((row) => (
-          <TableRow key={row.date}>
+        {externalMaterials.map((row, index) => (
+          <TableRow key={index}>
             <TableCell>{row.date}</TableCell>
             <TableCell>
-              <Link href={`https://campus.kpi.ua/student/index.php?mode=mob&show&irid=${row.resourceId}`}>
-                {row.name}
-              </Link>
+              <Link href={row.url}>{row.name}</Link>
             </TableCell>
             <TableCell className="max-w-[360px]">
               <div className="flex items-center gap-3">
